+++ include/roq/modify_order.h
@@ -23,13 +23,13 @@ namespace roq {
 
 //! Fields required to modify an existing order
 struct ROQ_PUBLIC ModifyOrder final {
-  std::string_view account;             //!< Account name
-  uint32_t order_id = {};               //!< Order identifier
-  double quantity = NaN;                //!< New (total) quantity
-  double price = NaN;                   //!< New limit price
-  std::string_view routing_id;          //!< Routing identifier
-  uint8_t version = {};                 //!< Version number (strictly increasing, optional)
-  uint8_t conditional_on_version = {};  //!< Auto-reject if this version has positively failed (optional)
+  std::string_view account;              //!< Account name
+  uint32_t order_id = {};                //!< Order identifier
+  double quantity = NaN;                 //!< New (total) quantity
+  double price = NaN;                    //!< New limit price
+  std::string_view routing_id;           //!< Routing identifier
+  uint32_t version = {};                 //!< Version number (strictly increasing, optional)
+  uint32_t conditional_on_version = {};  //!< Auto-reject if this version has positively failed (optional)
 };
 
 }  // namespace roq